add multi-backend demo
